+++ packages/app-lixi-api/src/modules/bookmark/bookmark.module.ts
@@ -4,11 +4,21 @@ import { BookmarkCacheService } from './bookmark-cache.service';
 import { BookmarkResolver } from './bookmark.resolver';
 import { PostCacheService } from '../page/post-cache.service';
 import { TimelineItemService } from '../timeline/timeline-item.service';
+import { EscrowOrderCacheService } from '../escrow/escrow-order/escrow-order-cache.service';
+import { DisputeCacheService } from '../escrow/dispute/dispute-cache.service';
 
 @Module({
   imports: [AuthModule],
   controllers: [],
-  providers: [BookmarkCacheService, BookmarkResolver, Logger, PostCacheService, TimelineItemService],
+  providers: [
+    BookmarkCacheService,
+    BookmarkResolver,
+    Logger,
+    PostCacheService,
+    TimelineItemService,
+    EscrowOrderCacheService,
+    DisputeCacheService
+  ],
   exports: [BookmarkCacheService, BookmarkResolver, Logger]
 })
 export class BookmarkModule {}