+++ packages/app-lixi-api/src/modules/escrow/dispute/dispute.resolver.ts
@@ -1,32 +1,35 @@
 import {
   Account,
+  BasicPaginationArgs,
   CreateDisputeInput,
   Dispute,
-  DisputeConnection,
-  DisputeOrder,
   DisputeStatus,
   EscrowOrderStatus,
-  PaginationArgs,
+  IBasicPaginated,
+  TIMELINE_TYPE,
+  TimelineItem,
+  TimelineItemConnection,
   UpdateDisputeInput
 } from '@bcpros/lixi-models';
-import { findManyCursorConnection } from '@devoxa/prisma-relay-cursor-connection';
 import { Logger, UseFilters, UseGuards } from '@nestjs/common';
-import { Args, Mutation, Parent, Query, ResolveField, Resolver, Subscription } from '@nestjs/graphql';
+import { Args, Mutation, Parent, Query, ResolveField, Resolver } from '@nestjs/graphql';
 import { SkipThrottle } from '@nestjs/throttler';
-import { PubSub } from 'graphql-subscriptions';
 import * as _ from 'lodash';
 import { I18n, I18nService } from 'nestjs-i18n';
 import { GqlHttpExceptionFilter } from 'src/middlewares/gql.exception.filter';
-import { PrismaService } from '../prisma/prisma.service';
-import { GqlJwtAuthGuard } from '../auth/guards/gql-jwtauth.guard';
+import { PrismaService } from '../../prisma/prisma.service';
+import { GqlJwtAuthGuard } from '../../auth/guards/gql-jwtauth.guard';
 import { AccountEntity } from 'src/decorators';
+import DisputeLoader from './dispute.loader';
+import { VError } from 'verror';
+import { DisputeCacheService } from './dispute-cache.service';
+import { TimelineItemService } from 'src/modules/timeline/timeline-item.service';
+import { createEdge } from 'src/common/custom-graphql-relay/paginate';
 import { InjectBot } from 'nestjs-telegraf';
-import { TELEGRAM_LOCAL_ECASH_BOT_NAME } from '../telegram/telegram-bot.constants';
+import { TELEGRAM_LOCAL_ECASH_BOT_NAME } from '../../telegram/telegram-bot.constants';
 import { Context, Telegraf } from 'telegraf';
 import { format } from 'node:util';
 import { BOT } from 'src/utils/bot.constants';
-
-const pubSub = new PubSub();
 
 @SkipThrottle()
 @Resolver(() => Dispute)
@@ -36,13 +39,11 @@ export class DisputeResolver {
     private logger: Logger,
     private prisma: PrismaService,
     @I18n() private i18n: I18nService,
+    private readonly disputeLoader: DisputeLoader,
+    private readonly disputeCacheService: DisputeCacheService,
+    private readonly timelineItemService: TimelineItemService,
     @InjectBot(TELEGRAM_LOCAL_ECASH_BOT_NAME) private bot: Telegraf<Context>
   ) {}
-
-  @Subscription(() => Dispute)
-  disputeCreated() {
-    return pubSub.asyncIterator('disputeCreated');
-  }
 
   @Query(() => Dispute)
   @UseGuards(GqlJwtAuthGuard)
@@ -68,40 +69,30 @@ export class DisputeResolver {
     return result;
   }
 
-  @Query(() => DisputeConnection)
+  @Query(() => TimelineItemConnection)
   @UseGuards(GqlJwtAuthGuard)
-  async allDisputesByAccountId(
-    @Args() { after, before, first, last }: PaginationArgs,
+  async allDisputeByAccount(
     @AccountEntity() account: Account,
-    @Args({
-      name: 'orderBy',
-      type: () => DisputeOrder,
-      nullable: true
-    })
-    orderBy: DisputeOrder
+    @Args() { after, first }: BasicPaginationArgs,
+    @Args({ name: 'disputeStatus', type: () => DisputeStatus }) disputeStatus: DisputeStatus
   ) {
-    const result = await findManyCursorConnection(
-      args =>
-        this.prisma.dispute.findMany({
-          where: {
-            escrowOrder: {
-              OR: [
-                {
-                  arbitratorAccountId: account.id
-                },
-                {
-                  moderatorAccountId: account.id
-                }
-              ]
-            }
-          },
-          include: { escrowOrder: true },
-          orderBy: orderBy ? { [orderBy.field]: orderBy.direction } : undefined,
-          ...args
-        }),
-      () => this.prisma.dispute.count({}),
-      { first, last, before, after }
+    if (!account) {
+      const accountNotExistMessage = await this.i18n.t('account.messages.accountNotExist');
+      throw new VError(accountNotExistMessage);
+    }
+    const paginated = await this.disputeCacheService.getPaginatedMyDisputeTimelineByTime(
+      account.id,
+      disputeStatus,
+      first,
+      after
     );
+    const timelineIds = paginated.edges.map(item => item.cursor);
+    // const timelines = await this.escrowOrderCacheService.getByIds(timelineIds);
+    const timelines = await this.timelineItemService.getByIds(timelineIds, TIMELINE_TYPE.DISPUTE);
+    const result = {
+      ...paginated,
+      edges: timelines.map(timeline => (timeline ? createEdge<TimelineItem>(timeline, 'id') : null))
+    } as IBasicPaginated<TimelineItem>;
     return result;
   }
 
@@ -151,7 +142,6 @@ export class DisputeResolver {
           }
         }
       });
-
       const url = `${process.env.LOCAL_ECASH_URL}/order-detail?id=${escrowOrder.id}`;
 
       if (createdBy === buyerAccount.publicKey && sellerAccount.telegramId) {
@@ -168,7 +158,6 @@ export class DisputeResolver {
         });
       }
 
-      pubSub.publish('disputeCreated', { disputeCreated: dispute });
       return dispute;
     } catch (e) {
       this.logger.log(e);
@@ -203,19 +192,11 @@ export class DisputeResolver {
       }
     });
 
-    pubSub.publish('disputeCreated', { disputeCreated: dispute });
     return dispute;
   }
 
   @ResolveField()
   async escrowOrder(@Parent() dispute: Dispute) {
-    const escrowOrder = this.prisma.escrowOrder.findFirst({
-      where: {
-        dispute: {
-          id: dispute.id
-        }
-      }
-    });
-    return escrowOrder;
+    return this.disputeLoader.batchEscrowOrders.load(dispute.escrowOrderId);
   }
 }