+++ packages/redux-store/src/store/post/poll.generated.ts
@@ -48,6 +48,7 @@ export type CreatePollMutation = {
       name: string;
       avatar?: string | null;
       createCommentFee?: string | null;
+      telegramUsername?: string | null;
     };
     page?: {
       __typename?: 'Page';
@@ -118,7 +119,7 @@ export type CreatePollMutation = {
         }> | null;
       }>;
     } | null;
-    offer?: {
+    postOffer?: {
       __typename?: 'Offer';
       postId: string;
       publicKey: string;